Changes done as per your response
