+++ components/allEvents.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useSession } from "next-auth/react";
 import Cookies from "js-cookie";
 import Image from "next/image";
+import Feedback from "./feedback";
 
 const AllEvents = () => {
   const [events, setEvents] = useState([]);
@@ -10,16 +11,12 @@ const AllEvents = () => {
   const userId = Cookies.get("_id");
   const [selectedEvent, setSelectedEvent] = useState(null);
   const [showFeedbackForm, setShowFeedbackForm] = useState(false);
-  const [rating, setRating] = useState("");
-  const [detailedFeedback, setDetailedFeedback] = useState("");
   let Allset = userId !== undefined;
 
   useEffect(() => {
-    // Fetch events from the API
     fetchEvents();
   }, []);
 
-  // Function to fetch events data
   const fetchEvents = () => {
     fetch("/api/auth/fetchevents")
       .then((response) => response.json())
@@ -34,7 +31,6 @@ const AllEvents = () => {
       });
   };
 
-  // Function to check if the user is registered for an event
   const checkRegister = (eventRegistered) => {
     for (let i = 0; i < eventRegistered.length; i++) {
       for (let j = 0; j < eventRegistered[i].length; j++) {
@@ -47,7 +43,6 @@ const AllEvents = () => {
     return false;
   };
 
-  // Function to handle event registration
   const handleRegister = async (event) => {
     try {
       const response = await fetch(
@@ -61,38 +56,6 @@ const AllEvents = () => {
       }
     } catch (error) {
       console.error("Error while registering:", error);
-    }
-  };
-
-  // Function to handle feedback submission
-  const handleFeedback = async () => {
-    try {
-      // Ensure both rating and detailed feedback are provided
-      if (!rating || !detailedFeedback) {
-        console.log("Please provide both rating and detailed feedback.");
-        return;
-      }
-
-      // Feedback data to be submitted
-      const feedbackData = {
-        eventId: selectedEvent._id,
-        userId: userId,
-        rating: rating,
-        detailedFeedback: detailedFeedback,
-      };
-
-      // Send POST request to submit feedback
-      // You need to implement the server-side logic for this part
-      console.log("Feedback data:", feedbackData);
-
-      // Close the feedback form
-      setShowFeedbackForm(false);
-
-      // Reset the feedback form state
-      setRating("");
-      setDetailedFeedback("");
-    } catch (error) {
-      console.error("Error while submitting feedback:", error);
     }
   };
 
@@ -194,51 +157,19 @@ const AllEvents = () => {
                   Login and fill data in settings to Register
                 </div>
               )}
+
+              {/* Feedback Form */}
+              {showFeedbackForm && (
+                <Feedback
+                  event={selectedEvent}
+                  userId={userId}
+                  setShowFeedbackForm={setShowFeedbackForm}
+                />
+              )}
             </div>
           ))
         )}
       </div>
-
-      {/* Feedback Form */}
-      {showFeedbackForm && (
-        <div className="fixed inset-0 bg-gray-700 bg-opacity-75 flex justify-center items-center">
-          <div className="bg-white p-6 rounded-md">
-            <h2 className="text-2xl font-bold mb-4">
-              Feedback Form {/*for {event.title}*/}
-            </h2>
-            <div className="mb-4">
-              <label className="block text-sm font-medium text-gray-700">
-                Rating:
-              </label>
-              <input
-                type="number"
-                min="1"
-                max="5"
-                placeholder="Enter rating (1-5)"
-                className="mt-1 p-2 border rounded-md w-full"
-                value={rating}
-                onChange={(e) => setRating(e.target.value)}
-              />
-            </div>
-            <div className="mb-4">
-              <label className="block text-sm font-medium text-gray-700">
-                Detailed Feedback:
-              </label>
-              <textarea
-                className="mt-1 p-2 border rounded-md w-full"
-                value={detailedFeedback}
-                onChange={(e) => setDetailedFeedback(e.target.value)}
-              ></textarea>
-            </div>
-            <button
-              className="px-4 py-2 bg-blue-500 text-white rounded-md hover:bg-blue-600"
-              onClick={handleFeedback}
-            >
-              Submit Feedback
-            </button>
-          </div>
-        </div>
-      )}
     </main>
   );
 };
